+++ novae/__init__.py
@@ -10,6 +10,7 @@ from . import utils
 from . import data
 from . import monitor
 from . import plot
+from .utils import spatial_neighbors, batch_effect_correction
 
 log = logging.getLogger("novae")
 configure_logger(log)