+++ novae/plot/_spatial.py
@@ -11,7 +11,11 @@ from scanpy._utils import sanitize_anndata
 
 from .. import utils
 from .._constants import Keys
-from ._utils import _subplots_per_slide, get_categorical_color_palette
+from ._utils import (
+    _get_default_cell_size,
+    _subplots_per_slide,
+    get_categorical_color_palette,
+)
 
 log = logging.getLogger(__name__)
 
@@ -20,7 +24,7 @@ def domains(
     adata: AnnData | list[AnnData],
     obs_key: str | None = None,
     slide_name_key: str | None = None,
-    cell_size: int | None = 10,
+    cell_size: int | None = None,
     ncols: int = 4,
     fig_size_per_slide: tuple[int, int] = (5, 5),
     na_color: str = "#ccc",
@@ -37,7 +41,7 @@ def domains(
         adata: An `AnnData` object, or a list of `AnnData` objects.
         obs_key: Name of the key from `adata.obs` containing the Novae domains. By default, the last available domain key is shown.
         slide_name_key: Key of `adata.obs` that contains the slide names. By default, uses the Novae unique slide ID.
-        cell_size: Size of the cells or spots.
+        cell_size: Size of the cells or spots. By default, it is calculated based on the median distance between cells.
         ncols: Number of columns to be shown.
         fig_size_per_slide: Size of the figure for each slide.
         na_color: Color for cells that does not belong to any domain (i.e. cells with a too small neighborhood).
@@ -56,6 +60,7 @@ def domains(
         sanitize_anndata(adata)
 
     all_domains, colors = get_categorical_color_palette(adatas, obs_key)
+    cell_size = cell_size or _get_default_cell_size(adata)
 
     fig, axes = _subplots_per_slide(adatas, ncols, fig_size_per_slide)
 
@@ -108,7 +113,7 @@ def spatially_variable_genes(
     adata: AnnData,
     obs_key: str | None = None,
     top_k: int = 5,
-    cell_size: int = 10,
+    cell_size: int | None = None,
     min_positive_ratio: float = 0.05,
     return_list: bool = False,
     show: bool = True,
@@ -123,7 +128,7 @@ def spatially_variable_genes(
         adata: An `AnnData` object corresponding to one slide.
         obs_key: Key in `adata.obs` that contains the domains. By default, it will use the last available Novae domain key.
         top_k: Number of SVG to be shown.
-        cell_size: Size of the cells or spots (`spot_size` argument of `sc.pl.spatial`).
+        cell_size: Size of the cells or spots (`spot_size` argument of `sc.pl.spatial`). By default, it is calculated based on the median distance between cells.
         min_positive_ratio: Genes whose "ratio of cells expressing it" is lower than this threshold are not considered.
         return_list: Whether to return the list of SVG instead of plotting them.
         show: Whether to show the plot.
@@ -156,4 +161,4 @@ def spatially_variable_genes(
     if return_list:
         return svg.tolist()
 
-    sc.pl.spatial(adata, color=svg, spot_size=cell_size, show=show, **kwargs)
+    sc.pl.spatial(adata, color=svg, spot_size=cell_size or _get_default_cell_size(adata), show=show, **kwargs)