+++ novae/plot/_utils.py
@@ -32,3 +32,15 @@ def _subplots_per_slide(
     )
 
     return fig, axes
+
+
+def _get_default_cell_size(adata: AnnData | list[AnnData]) -> float:
+    if isinstance(adata, list):
+        adata = max(adata, key=lambda adata: adata.n_obs)
+
+    assert (
+        Keys.ADJ in adata.obsp
+    ), f"Expected {Keys.ADJ} in adata.obsp. Please run `novae.spatial_neighbors(...)` first."
+
+    median_distance = np.median(adata.obsp[Keys.ADJ].data)
+    return median_distance * 0.8